Move DeletedMessageData next to DeletedMessage
